add docker bridge network support
