+++ bcs/contract/native/config_darwin.go
@@ -1,0 +1,6 @@
+package native
+
+const (
+	chainAddrDocker = "docker.for.mac.host.internal"
+	chainAddrHost   = "localhost"
+)
