+++ bcs/contract/native/config_linux.go
@@ -1,0 +1,6 @@
+package native
+
+const (
+	chainAddrDocker = "localhost"
+	chainAddrHost   = "localhost"
+)
