+++ bcs/contract/native/testdata/go.mod
@@ -2,4 +2,4 @@ module contract-test
 
 go 1.15
 
-require github.com/xuperchain/contract-sdk-go v0.0.0-20210118130759-2897ea0f985a
+require github.com/xuperchain/contract-sdk-go v1.0.1-0.20220608031628-2a571766e240
